currentCoupon migrated
